Added an Enemy
What is New:
-Added an Enemy With animations and a model
- Gave the enemy Stamina, ability to attack, and chase the player
-Player Can Die now from the enemy

Enemy doesn't have health as of now and player doesn't have a way to kill the enemy but it'll be easy to implement. I will probably add a spawner of sorts in the future and a way to kill off these dudes but basically for the most part it is done.